% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{add_to_log}
\alias{add_to_log}
\title{Add item to log}
\usage{
add_to_log(
  object,
  message = "",
  time = Sys.time(),
  level = "INFO",
  print = FALSE
)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
Add item to log
}
