% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlascaModel.R
\name{AlascaModel}
\alias{AlascaModel}
\title{R6 Class Representing an ALASCA model}
\description{
The object builds the ALASCA model and contains the data
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{df}}{Data table/frame. The data to analyze}

\item{\code{formula}}{An AlascaForula object}

\item{\code{wide}}{Boolean. Whether the provided data is in wide format}

\item{\code{scale_function}}{How to scale the data. Options are \code{NULL}, custom function, or \code{"sdall"}, \code{"sdref"}, \code{"sdt1"}, \code{"sdreft1"}}

\item{\code{ignore_missing}}{If TRUE, ignore missing predictive values}

\item{\code{ignore_missing_covars}}{If TRUE, ignore missing covariate values}

\item{\code{version}}{Version number}

\item{\code{update_date}}{Date of latest update}

\item{\code{separate_effects}}{If TRUE, try to separate the effects}

\item{\code{equal_baseline}}{If TRUE, remove interaction between baselines}

\item{\code{effect_list}}{List. Contains info related to the effects}

\item{\code{n_validation_folds}}{Integer. If using jack-knife validation, exclude 1/n_validation_folds of the participants at each iteration}

\item{\code{n_validation_runs}}{Integer. Number of iterations to use for validation}

\item{\code{validation_quantile_method}}{Integer between 1 and 9. See \code{\link[stats:quantile]{stats::quantile()}} for details}

\item{\code{save_validation_ids}}{If TRUE, save the participants in each validation iteration to a csv file}

\item{\code{optimize_score}}{If TRUE, test all combinations of signs for the most important loadings, and choose the combination being the best fit}

\item{\code{validate}}{If TRUE, validate the model}

\item{\code{validate_regression}}{If TRUE, validate get marginal means}

\item{\code{validation}}{Boolean. Synonym to \code{validate}}

\item{\code{validation_method}}{String. Defines the validation method; \code{"bootstrap"} (default) or \code{"jack-knife"}}

\item{\code{validation_ids}}{A data frame where each row contains the ids for one validation iteration}

\item{\code{validation_assign_new_ids}}{Logical. Assign new IDs during validation. Must be \code{TRUE} for reduce_dimensions to work}

\item{\code{limitsCI}}{Lower and upper quantile to use for validation}

\item{\code{compress_validation}}{Integer between 0 and 100. See \code{\link[fst:write_fst]{fst::write_fst()}} for details}

\item{\code{reduce_dimensions}}{Boolean. Use PCA to reduce data dimensions prior to analysis}

\item{\code{pca_function}}{String or custom function. Which pca function to use for dimension reduction, either "prcomp" or "irlba" or "princomp" or custom function}

\item{\code{save_to_disk}}{Write model data to disk to reduce memory usage}

\item{\code{db_method}}{String. Use a \code{"duckdb"} or a \code{"SQLite"} database for validation data}

\item{\code{filename}}{Filename for the saved model}

\item{\code{filepath}}{Where to save the model. Defaults to \verb{ALASCA/<timestamp>}}

\item{\code{save}}{Save model data and plots}

\item{\code{method}}{String. Can be \code{"LM"} or \code{"LMM"}}

\item{\code{max_PC}}{Integer. The maximal number of principal components to keep for further analysis}

\item{\code{use_Rfast}}{Boolean. If \code{TRUE} (default), use Rfast, else use lm or lme4}

\item{\code{p_adjust_method}}{String. See \code{\link[stats:p.adjust]{stats::p.adjust()}}}

\item{\code{participant_column}}{String. The column used for IDs. If not provided, it will guess based on random effect or \code{ID}}

\item{\code{stratification_column}}{String. Name of the column to use for stratification during validation}

\item{\code{explanatory_limit}}{Only validate components explaining more than \code{explanatory_limit} of the variance}

\item{\code{init_time}}{The time when the object is initialized}

\item{\code{log_to}}{String deciding logging target: \code{"all"} (default), \code{"file"}, \code{"console"}, \code{"none"}}

\item{\code{log_level}}{String. What level of log messages to print; \code{"DEBUG"}, \code{"INFO"}, \code{"WARN"}, \code{"ERROR"}}

\item{\code{do_debug}}{Boolean. Log more details}

\item{\code{finished}}{Boolean. Indicates whether the model has been successfully initiated}

\item{\code{ALASCA}}{List. Contains all model outputs: \code{score}, \code{loading}, \code{explained} and \code{significant_PCs}}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{get_plot_group}}{Name of the grouping factor (used for plotting)}

\item{\code{effect_terms}}{List of the terms in the effect matrices}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-remove_embedded_data}{\code{AlascaModel$remove_embedded_data()}}
\item \href{#method-get_default_scaling_function}{\code{AlascaModel$get_default_scaling_function()}}
\item \href{#method-get_scaling_function}{\code{AlascaModel$get_scaling_function()}}
\item \href{#method-get_pca_function}{\code{AlascaModel$get_pca_function()}}
\item \href{#method-build_model}{\code{AlascaModel$build_model()}}
\item \href{#method-run_regression}{\code{AlascaModel$run_regression()}}
\item \href{#method-get_regression_coefficients}{\code{AlascaModel$get_regression_coefficients()}}
\item \href{#method-remove_covars}{\code{AlascaModel$remove_covars()}}
\item \href{#method-get_effect_matrix}{\code{AlascaModel$get_effect_matrix()}}
\item \href{#method-do_pca}{\code{AlascaModel$do_pca()}}
\item \href{#method-do_reduce_dimensions}{\code{AlascaModel$do_reduce_dimensions()}}
\item \href{#method-clean_pca}{\code{AlascaModel$clean_pca()}}
\item \href{#method-clean_alasca}{\code{AlascaModel$clean_alasca()}}
\item \href{#method-do_validate}{\code{AlascaModel$do_validate()}}
\item \href{#method-get_validation_percentiles}{\code{AlascaModel$get_validation_percentiles()}}
\item \href{#method-get_validation_percentiles_regression}{\code{AlascaModel$get_validation_percentiles_regression()}}
\item \href{#method-get_validation_percentiles_covars}{\code{AlascaModel$get_validation_percentiles_covars()}}
\item \href{#method-get_validation_percentiles_loading}{\code{AlascaModel$get_validation_percentiles_loading()}}
\item \href{#method-get_validation_scores}{\code{AlascaModel$get_validation_scores()}}
\item \href{#method-get_validation_loadings}{\code{AlascaModel$get_validation_loadings()}}
\item \href{#method-get_validation_percentiles_score}{\code{AlascaModel$get_validation_percentiles_score()}}
\item \href{#method-get_regression_predictions}{\code{AlascaModel$get_regression_predictions()}}
\item \href{#method-get_validation_ids}{\code{AlascaModel$get_validation_ids()}}
\item \href{#method-prepare_validation_run}{\code{AlascaModel$prepare_validation_run()}}
\item \href{#method-get_bootstrap_data}{\code{AlascaModel$get_bootstrap_data()}}
\item \href{#method-new}{\code{AlascaModel$new()}}
\item \href{#method-finalize}{\code{AlascaModel$finalize()}}
\item \href{#method-update}{\code{AlascaModel$update()}}
\item \href{#method-log}{\code{AlascaModel$log()}}
\item \href{#method-plot}{\code{AlascaModel$plot()}}
\item \href{#method-get_residuals}{\code{AlascaModel$get_residuals()}}
\item \href{#method-set_effect_terms}{\code{AlascaModel$set_effect_terms()}}
\item \href{#method-set_design_matrices}{\code{AlascaModel$set_design_matrices()}}
\item \href{#method-flip}{\code{AlascaModel$flip()}}
\item \href{#method-get_ref}{\code{AlascaModel$get_ref()}}
\item \href{#method-get_levels}{\code{AlascaModel$get_levels()}}
\item \href{#method-get_PCs}{\code{AlascaModel$get_PCs()}}
\item \href{#method-get_predictions}{\code{AlascaModel$get_predictions()}}
\item \href{#method-get_scores}{\code{AlascaModel$get_scores()}}
\item \href{#method-get_loadings}{\code{AlascaModel$get_loadings()}}
\item \href{#method-get_covars}{\code{AlascaModel$get_covars()}}
\item \href{#method-print_version}{\code{AlascaModel$print_version()}}
\item \href{#method-save_validation}{\code{AlascaModel$save_validation()}}
\item \href{#method-save_to_csv}{\code{AlascaModel$save_to_csv()}}
\item \href{#method-save_model}{\code{AlascaModel$save_model()}}
\item \href{#method-rotate_matrix_optimize_score}{\code{AlascaModel$rotate_matrix_optimize_score()}}
\item \href{#method-rotate_matrix}{\code{AlascaModel$rotate_matrix()}}
\item \href{#method-clone}{\code{AlascaModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_embedded_data"></a>}}
\if{latex}{\out{\hypertarget{method-remove_embedded_data}{}}}
\subsection{Method \code{remove_embedded_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$remove_embedded_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_default_scaling_function"></a>}}
\if{latex}{\out{\hypertarget{method-get_default_scaling_function}{}}}
\subsection{Method \code{get_default_scaling_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_default_scaling_function()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scaling_function"></a>}}
\if{latex}{\out{\hypertarget{method-get_scaling_function}{}}}
\subsection{Method \code{get_scaling_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_scaling_function()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pca_function"></a>}}
\if{latex}{\out{\hypertarget{method-get_pca_function}{}}}
\subsection{Method \code{get_pca_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_pca_function()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build_model"></a>}}
\if{latex}{\out{\hypertarget{method-build_model}{}}}
\subsection{Method \code{build_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$build_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_regression"></a>}}
\if{latex}{\out{\hypertarget{method-run_regression}{}}}
\subsection{Method \code{run_regression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$run_regression()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_regression_coefficients"></a>}}
\if{latex}{\out{\hypertarget{method-get_regression_coefficients}{}}}
\subsection{Method \code{get_regression_coefficients()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_regression_coefficients()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_covars"></a>}}
\if{latex}{\out{\hypertarget{method-remove_covars}{}}}
\subsection{Method \code{remove_covars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$remove_covars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_effect_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-get_effect_matrix}{}}}
\subsection{Method \code{get_effect_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_effect_matrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-do_pca"></a>}}
\if{latex}{\out{\hypertarget{method-do_pca}{}}}
\subsection{Method \code{do_pca()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$do_pca()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-do_reduce_dimensions"></a>}}
\if{latex}{\out{\hypertarget{method-do_reduce_dimensions}{}}}
\subsection{Method \code{do_reduce_dimensions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$do_reduce_dimensions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_pca"></a>}}
\if{latex}{\out{\hypertarget{method-clean_pca}{}}}
\subsection{Method \code{clean_pca()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$clean_pca()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_alasca"></a>}}
\if{latex}{\out{\hypertarget{method-clean_alasca}{}}}
\subsection{Method \code{clean_alasca()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$clean_alasca()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-do_validate"></a>}}
\if{latex}{\out{\hypertarget{method-do_validate}{}}}
\subsection{Method \code{do_validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$do_validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_validation_percentiles"></a>}}
\if{latex}{\out{\hypertarget{method-get_validation_percentiles}{}}}
\subsection{Method \code{get_validation_percentiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_validation_percentiles(objectlist)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_validation_percentiles_regression"></a>}}
\if{latex}{\out{\hypertarget{method-get_validation_percentiles_regression}{}}}
\subsection{Method \code{get_validation_percentiles_regression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_validation_percentiles_regression(objectlist)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_validation_percentiles_covars"></a>}}
\if{latex}{\out{\hypertarget{method-get_validation_percentiles_covars}{}}}
\subsection{Method \code{get_validation_percentiles_covars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_validation_percentiles_covars(objectlist)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_validation_percentiles_loading"></a>}}
\if{latex}{\out{\hypertarget{method-get_validation_percentiles_loading}{}}}
\subsection{Method \code{get_validation_percentiles_loading()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_validation_percentiles_loading(objectlist)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_validation_scores"></a>}}
\if{latex}{\out{\hypertarget{method-get_validation_scores}{}}}
\subsection{Method \code{get_validation_scores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_validation_scores(objectlist = NULL, effect_i = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_validation_loadings"></a>}}
\if{latex}{\out{\hypertarget{method-get_validation_loadings}{}}}
\subsection{Method \code{get_validation_loadings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_validation_loadings(objectlist = NULL, effect_i = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_validation_percentiles_score"></a>}}
\if{latex}{\out{\hypertarget{method-get_validation_percentiles_score}{}}}
\subsection{Method \code{get_validation_percentiles_score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_validation_percentiles_score(objectlist)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_regression_predictions"></a>}}
\if{latex}{\out{\hypertarget{method-get_regression_predictions}{}}}
\subsection{Method \code{get_regression_predictions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_regression_predictions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_validation_ids"></a>}}
\if{latex}{\out{\hypertarget{method-get_validation_ids}{}}}
\subsection{Method \code{get_validation_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_validation_ids()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare_validation_run"></a>}}
\if{latex}{\out{\hypertarget{method-prepare_validation_run}{}}}
\subsection{Method \code{prepare_validation_run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$prepare_validation_run(runN)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_bootstrap_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_bootstrap_data}{}}}
\subsection{Method \code{get_bootstrap_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_bootstrap_data(df_raw, participants_in_bootstrap, modmat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$new(df, formula, effects, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Update the current model (used for validation)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log"></a>}}
\if{latex}{\out{\hypertarget{method-log}{}}}
\subsection{Method \code{log()}}{
Function for logging messages using the log4r package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$log(message, level = "INFO")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{The message to log}

\item{\code{level}}{Level of the message; DEBUG, INFO, WARN, ERROR, FATAL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Main function for plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$plot(effect = 1, component = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{effect}}{Integer or vector. Which(s) effect(s) to plot}

\item{\code{component}}{Integer or vector. Which(s) component(s) to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_residuals"></a>}}
\if{latex}{\out{\hypertarget{method-get_residuals}{}}}
\subsection{Method \code{get_residuals()}}{
Resample participants for validation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_residuals(variable = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_effect_terms"></a>}}
\if{latex}{\out{\hypertarget{method-set_effect_terms}{}}}
\subsection{Method \code{set_effect_terms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$set_effect_terms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_design_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-set_design_matrices}{}}}
\subsection{Method \code{set_design_matrices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$set_design_matrices()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-flip"></a>}}
\if{latex}{\out{\hypertarget{method-flip}{}}}
\subsection{Method \code{flip()}}{
Switch the sign of scores and loadings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$flip(effect_i = 0, component = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{effect_i}}{The effect to reflect, \code{0} or \code{NULL} reflects the entire model}

\item{\code{component}}{The component to reflect, \code{0} or \code{NULL} reflects the entire model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ref"></a>}}
\if{latex}{\out{\hypertarget{method-get_ref}{}}}
\subsection{Method \code{get_ref()}}{
Returns the reference level of a given column
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_ref(columns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{A column containing factors}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The reference level
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_levels"></a>}}
\if{latex}{\out{\hypertarget{method-get_levels}{}}}
\subsection{Method \code{get_levels()}}{
Returns all the levels of a given column
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_levels(column, reduced = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column}}{A column containing factors}

\item{\code{reduced}}{Boolean. If \code{TRUE}, returns the PCs instead of actual variables if dimensions are reduced with PCA}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector with factor levels
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_PCs"></a>}}
\if{latex}{\out{\hypertarget{method-get_PCs}{}}}
\subsection{Method \code{get_PCs()}}{
Returns the most interesting principal components (i.e., components explaining more than a given limit of variance: \code{explanatory_limit})
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_PCs(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Index corresponding to the effect of interest}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector with principal components
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_predictions"></a>}}
\if{latex}{\out{\hypertarget{method-get_predictions}{}}}
\subsection{Method \code{get_predictions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_predictions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scores"></a>}}
\if{latex}{\out{\hypertarget{method-get_scores}{}}}
\subsection{Method \code{get_scores()}}{
Return scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_scores(effect_i = 1, component = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{effect_i}}{The effect to reflect, \code{0} or \code{NULL} reflects the entire model}

\item{\code{component}}{The component to reflect, \code{0} or \code{NULL} reflects the entire model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_loadings"></a>}}
\if{latex}{\out{\hypertarget{method-get_loadings}{}}}
\subsection{Method \code{get_loadings()}}{
Return loadings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_loadings(effect_i = 1, component = 1, n_limit = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{effect_i}}{The effect to reflect, \code{0} or \code{NULL} reflects the entire model}

\item{\code{component}}{The component to reflect, \code{0} or \code{NULL} reflects the entire model}

\item{\code{n_limit}}{Return only two times this number of loadings (the \code{n_limit} highest and lowest loadings). Use \code{0} to return all (default)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_covars"></a>}}
\if{latex}{\out{\hypertarget{method-get_covars}{}}}
\subsection{Method \code{get_covars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$get_covars(n_limit = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_limit}}{Return only two times this number of coefficients (the \code{n_limit} highest and lowest coefficients). Use \code{0} to return all (default)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_version"></a>}}
\if{latex}{\out{\hypertarget{method-print_version}{}}}
\subsection{Method \code{print_version()}}{
Print ALASCA version
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$print_version()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String with version number and date of latest update
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_validation"></a>}}
\if{latex}{\out{\hypertarget{method-save_validation}{}}}
\subsection{Method \code{save_validation()}}{
Write scores, loadings, covars and predictions from validation run to database and remove data from memory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$save_validation(ii)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ii}}{Number of the validation run}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_to_csv"></a>}}
\if{latex}{\out{\hypertarget{method-save_to_csv}{}}}
\subsection{Method \code{save_to_csv()}}{
Save scores, loading, covars, and predictions to csv files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$save_to_csv()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_model"></a>}}
\if{latex}{\out{\hypertarget{method-save_model}{}}}
\subsection{Method \code{save_model()}}{
Save model to RDS and scores, loading, covars, and predictions to csv files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$save_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rotate_matrix_optimize_score"></a>}}
\if{latex}{\out{\hypertarget{method-rotate_matrix_optimize_score}{}}}
\subsection{Method \code{rotate_matrix_optimize_score()}}{
Rotate model loadings and scores with procrustes. This function checks all combinations of signs to minimize variation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$rotate_matrix_optimize_score(target)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{Rotate model (\code{self}) with this as target}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rotate_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-rotate_matrix}{}}}
\subsection{Method \code{rotate_matrix()}}{
Rotate model loadings and scores with procrustes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$rotate_matrix(target)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{Rotate model (\code{self}) with this as target}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlascaModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
