% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{appendModel}
\alias{appendModel}
\title{Append a new model to an existing model}
\usage{
appendModel(target, object)
}
\arguments{
\item{target}{The existing model}

\item{object}{The new model}
}
\value{
An ALASCA object
}
\description{
Append a new model to an existing model
}
