% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{assessGroupDifferences}
\alias{assessGroupDifferences}
\title{Assess group differences}
\usage{
assessGroupDifferences(
  object,
  variables = NA,
  doPlot = TRUE,
  filetypes = "png",
  figsize = c(12, 8, 300),
  rawOut = FALSE
)
}
\arguments{
\item{object}{An ALASCA object}

\item{variables}{Return these variables, defaults to all}

\item{doPlot}{Set to \code{FALSE} to return data frames instead of plot}

\item{rawOut}{Set to \code{TRUE} to return emmeans plot or object}
}
\value{
A plot or an emmeans object
}
\description{
Assess group differences
}
