% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R, R/buildModel.R
\name{build_model}
\alias{build_model}
\title{Organize the ALASCA model construction}
\usage{
build_model()

build_model()
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object

An ALASCA object
}
\description{
This function builds the ALASCA model

This function builds the ALASCA model
}
