% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{check_that_columns_are_valid}
\alias{check_that_columns_are_valid}
\title{Check if columns are missing}
\usage{
check_that_columns_are_valid(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\description{
...
}
