% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{clean_pca}
\alias{clean_pca}
\title{Clean the PCA data}
\usage{
clean_pca(object)
}
\arguments{
\item{object}{An ALASCA object to be sanitized}
}
\value{
An ALASCA object
}
\description{
This function makes the pca output more useful
}
