% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{cleanPCA}
\alias{cleanPCA}
\title{Clean the PCA data}
\usage{
cleanPCA(object)
}
\arguments{
\item{object}{An RMASCA object to be sanitized}
}
\value{
An RMASCA object
}
\description{
This function makes the pca output more useful
}
