% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{count_participants}
\alias{count_participants}
\title{Count participants and samples}
\usage{
count_participants(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
A list of data frames
}
\description{
Count participants and samples
}
