% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R, R/pca.R
\name{do_pca}
\alias{do_pca}
\title{Perform PCA}
\usage{
do_pca()

do_pca()
}
\arguments{
\item{object}{An ALASCA object to be sanitized}
}
\value{
An ALASCA object

An ALASCA object
}
\description{
This function performs PCA

This function performs PCA
}
