% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{do_validate}
\alias{do_validate}
\title{Validate the ALASCA model LMM}
\usage{
do_validate()
}
\arguments{
\item{object}{An ALASCA object}

\item{participant_column}{The name of the column containing participant identifier. Needed if not set during initialization of the model.}

\item{validate_regression}{Whether to validate regression models}
}
\value{
An ALASCA object
}
\description{
This function performs leave-one-out robustness testing of your ALASCA model. If you didn't specify the number of runs \code{n_validation_runs} when initializing the model (see \code{\link{ALASCA}}), you can do it by running for example \code{model$n_validation_runs <- 100} prior to calling \code{validate}. Your dataset is divided into \code{n_validation_folds} partitions, keeping group proportions, and one of these are left out. \code{n_validation_folds} is set the same way as  \code{n_validation_runs}.
}
\examples{
load("PE.Rdata")
model$n_validation_runs <- 10
model.val <- validate(model, participant_column = "ID")
}
