% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{doLimmPCA}
\alias{doLimmPCA}
\title{Perform PCA for Limm-PCA}
\usage{
doLimmPCA(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function performs PCA before "the real PCA"
}
