% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.get_explained_label}
\alias{.get_explained_label}
\title{Get exploratory power for plot label}
\usage{
.get_explained_label(object, component = 1, effect = "time", type = "Score")
}
\arguments{
\item{object}{An ALASCA object}

\item{comp}{Which two components to plot (default: \verb{c(1, 2})}
}
\value{
A ggplot2 objects.
}
\description{
This function returns ...
}
