% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.getPlotHandle}
\alias{.getPlotHandle}
\title{Get plotting features}
\usage{
.getPlotHandle(
  g,
  object,
  myTheme,
  effect = "time",
  comps = c(1, 2),
  legend = NA
)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
A list with colors
}
\description{
This function returns a list with colors for plotting
}
