% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{find_missing_response_variables}
\alias{find_missing_response_variables}
\title{Check if response variables are missing}
\usage{
find_missing_response_variables()
}
\value{
An ALASCA object
}
\description{
Called from AlascaDataset
}
