% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{flip.AlascaModel}
\alias{flip.AlascaModel}
\title{Flip an ALASCA object}
\usage{
\method{flip}{AlascaModel}(object, effect = 0, component = 0)
}
\arguments{
\item{object}{An ALASCA object}

\item{effect}{Specify effect(s) to be flipped, use \code{0} or \code{NULL} to flip all (default)}

\item{component}{Component(s) to be flipped, use \code{0} or \code{NULL} to flip all (default)}
}
\value{
The rotated object
}
\description{
Changes the sign of loadings and scores
}
