% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{flip}
\alias{flip}
\title{Flip an ALASCA object}
\usage{
flip(object, component = NA, effect = "both")
}
\arguments{
\item{object}{An ALASCA object}

\item{component}{Components to be flipped, \code{NA} flips all (default)}

\item{effect}{Specify \code{time} or \code{group} to only flip subplot}
}
\value{
An ALASCA object
}
\description{
Changes the sign of loadings and scores
}
