% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{flipIt}
\alias{flipIt}
\title{Flip an ALASCA object}
\usage{
flipIt(object, component = NA)
}
\arguments{
\item{object}{An ALASCA object}

\item{component}{Components to be flipped, \code{NA} flips all (default)}
}
\value{
An ALASCA object
}
\description{
Changes the sign of loadings and scores
}
