% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_bootstrap_data}
\alias{get_bootstrap_data}
\title{Make bootstrap data set}
\usage{
get_bootstrap_data(df_raw, participants_in_bootstrap, modmat)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
A data frame
}
\description{
Get data frame
}
