% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_bootstrap_ids}
\alias{get_bootstrap_ids}
\title{Make bootstrap sample}
\usage{
get_bootstrap_ids(runN = NA)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
A data frame
}
\description{
Get data frame with new and old IDs
}
