% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_covars.AlascaModel}
\alias{get_covars.AlascaModel}
\title{Get coefficients for covariates}
\usage{
get_covars.AlascaModel(object, n_limit = 0)
}
\arguments{
\item{object}{An ALASCA object}

\item{n_limit}{Return only the \code{n_limit} largest and smallest coefficients}
}
\value{
A data table
}
\description{
Get coefficients for covariates
}
