% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_covars}
\alias{get_covars}
\title{Get covariables}
\usage{
get_covars(object, n.limit = 0)
}
\arguments{
\item{object}{An ALASCA object}

\item{n.limit}{Returns the n highest and lowest loadings by PC (i.e., 2*n.limit loadings per PC)}
}
\value{
A list with scores for time (and group), and the exploratory power for each component
}
\description{
This function returns the other covariables in an ALASCA model
}
