% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_info_from_formula}
\alias{get_info_from_formula}
\title{Get information from formula}
\usage{
get_info_from_formula(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
...
}
