% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_loading_plot}
\alias{get_loading_plot}
\title{Get loading plot}
\usage{
get_loading_plot(
  object,
  component = 1,
  effect = "time",
  decreasing_loadings = TRUE,
  too_dense = NA,
  n_limit = 0,
  highlight = NA,
  filetype = NA,
  flip_axes = TRUE,
  plot_zeroline = TRUE,
  figsize = NA,
  figunit = NA,
  variables = NA,
  loadinggroup = NA,
  limit_loading = FALSE,
  sort_by_loadinggroup = TRUE,
  point_size = 0.4,
  my_theme = NA
)
}
\arguments{
\item{object}{An ALASCA object}

\item{component}{Which component to plot?}

\item{effect}{Plot time or group}

\item{decreasing_loadings}{Logical. Should loadings be sorted in decreasing order?}

\item{filetype}{Which filetype you want to save the figure to}

\item{flip_axes}{When \code{TRUE} (default), list the variable loadings vertical instead of horizontal}

\item{plot_zeroline}{When \code{TRUE} (default), plot a zero line in the loading plot}

\item{figsize}{A vector containing \code{c(widht,height,dpi)} (default: \code{c(12, 8, 300)})}

\item{my_theme}{A ggplot2 theme to use, defaults to \code{ggplot2::theme_bw()}}
}
\value{
A ggplot object
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
