% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_loadings.AlascaModel}
\alias{get_loadings.AlascaModel}
\title{Returns model loadings}
\usage{
\method{get_loadings}{AlascaModel}(object, effect = 1, component = 1, n_limit = 0)
}
\arguments{
\item{object}{An ALASCA object.}

\item{effect}{The number of the effect(s) of interest}

\item{component}{The principal component(s) of interest}

\item{n_limit}{Limit the number of variables returned}
}
\value{
A data table with loadings
}
\description{
Returns model loadings
}
