% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_loadings}
\alias{get_loadings}
\title{Get loadings}
\usage{
get_loadings(object, limit_loading = FALSE, n_limit = 0L, component = c(0))
}
\arguments{
\item{object}{An ALASCA object}

\item{n_limit}{Returns the n highest and lowest loadings by PC (i.e., 2*n_limit loadings per PC)}
}
\value{
A list with loadings for time (and group), and the exploratory power for each component
}
\description{
This function  returns the loadings for an ALASCA model
}
