% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_linetypes}
\alias{get_plot_linetypes}
\title{Get linetypes}
\usage{
get_plot_linetypes(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
A list with linetypes
}
\description{
This function returns a list with linetypes for plotting
}
