% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_palette}
\alias{get_plot_palette}
\title{Get color palette}
\usage{
get_plot_palette(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
A list with colors
}
\description{
This function returns a list with colors for plotting
}
