% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_predictions.AlascaModel}
\alias{get_predictions.AlascaModel}
\title{Returns model predictions}
\usage{
\method{get_predictions}{AlascaModel}(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
A data table with model predictions
}
\description{
Returns model predictions
}
