% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_predictions}
\alias{get_predictions}
\title{Get coefficients for covariates}
\usage{
get_predictions(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
A data table with regression coefficients
}
\description{
Get coefficients for covariates
}
