% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_regression_coefficients}
\alias{get_regression_coefficients}
\title{Get regression coefficients}
\usage{
get_regression_coefficients()
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function extract the regression coefficients for the ALASCA model
}
