% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_regression_formula}
\alias{get_regression_formula}
\title{Modify formula}
\usage{
get_regression_formula()
}
\value{
void
}
\description{
The regression formula must be adapted to the regression method
}
