% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_regression_predictions_standard}
\alias{get_regression_predictions_standard}
\title{Validate underlying regression models}
\usage{
get_regression_predictions_standard()
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function calculates predictions from each regression model
}
