% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{get_relevant_pcs}
\alias{get_relevant_pcs}
\title{Get relevant PCs}
\usage{
get_relevant_pcs(object, effect = "time")
}
\arguments{
\item{x}{Explanatory power of PC}
}
\value{
A vector with relevant PCs
}
\description{
This function extract percentiles during validation
}
