% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_scaling_function}
\alias{get_scaling_function}
\title{Get a scaling function}
\usage{
get_scaling_function(scaleFun_string = "sdall", scaleFun.center = TRUE)
}
\arguments{
\item{scaleFun_string}{String to define scaing function: \code{sdall}, \code{sdref}, \code{sdt1}, \code{sdreft1}}

\item{scaleFun.center}{Boolean. Mean centering}
}
\value{
An ALASCA object
}
\description{
Return scaling function
}
