% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_scores.AlascaModel}
\alias{get_scores.AlascaModel}
\title{Returns model scores}
\usage{
get_scores.AlascaModel(object, effect = 1, component = 1)
}
\arguments{
\item{object}{An AALSCA object.}

\item{effect}{The number of the effect of interest}

\item{component}{The principal component(s) of interest}
}
\description{
Returns model scores
}
