% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_scores}
\alias{get_scores}
\title{Returns model scores}
\usage{
get_scores(object, ...)
}
\arguments{
\item{object}{An ALASCA object.}
}
\value{
A data table with scores
}
\description{
Returns model scores
}
