% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{get_validation_percentiles}
\alias{get_validation_percentiles}
\title{Extract percentiles}
\usage{
get_validation_percentiles(object, objectlist)
}
\arguments{
\item{object}{ALASCA object}

\item{objectlist}{List of ALASCA objects}
}
\value{
An ALASCA object
}
\description{
This function extract percentiles during validation
}
