% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_validation_percentiles_covars}
\alias{get_validation_percentiles_covars}
\title{Extract percentiles for covariates}
\usage{
get_validation_percentiles_covars(objectlist)
}
\arguments{
\item{objectlist}{List of ALASCA objects}
}
\description{
This function extract percentiles for validation of covariates
}
