% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_validation_percentiles_loading}
\alias{get_validation_percentiles_loading}
\title{Extract percentiles for loading}
\usage{
get_validation_percentiles_loading(objectlist)
}
\arguments{
\item{objectlist}{List of ALASCA objects}
}
\description{
This function extract percentiles during validation of loadings
}
