% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_validation_percentiles_regression}
\alias{get_validation_percentiles_regression}
\title{Extract percentiles for regressions}
\usage{
get_validation_percentiles_regression(objectlist)
}
\arguments{
\item{objectlist}{List of ALASCA objects}
}
\value{
An ALASCA object
}
\description{
This function extract percentiles for validation of regression
}
