% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{get_validation_percentiles_score}
\alias{get_validation_percentiles_score}
\title{Extract percentiles for score}
\usage{
get_validation_percentiles_score(object, objectlist)
}
\arguments{
\item{object}{ALASCA object}

\item{objectlist}{List of ALASCA objects}
}
\value{
An ALASCA object
}
\description{
This function extract percentiles during validation of scores
}
