% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{getEffectMatrix}
\alias{getEffectMatrix}
\title{Get effect matrix}
\usage{
getEffectMatrix(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function calculates the effect matrix
}
