% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{getFilename}
\alias{getFilename}
\title{Get file path}
\usage{
getFilename(
  object,
  filename = NA,
  filepath = NA,
  prefix = "",
  suffix = "",
  filetype = ""
)
}
\arguments{
\item{object}{An ALASCA object}

\item{filename}{File name}

\item{filepath}{File path}

\item{prefix}{Prefix}

\item{suffix}{Suffix}
}
\value{
A ggplot2 objects.
}
\description{
Get file path
}
