% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMM.R
\name{getLMECoefficients}
\alias{getLMECoefficients}
\title{Get LMM coefficients}
\usage{
getLMECoefficients(object)
}
\arguments{
\item{object}{An RMASCA object to be sanitized}
}
\value{
An RMASCA object
}
\description{
This function calculates the LMM coefficients for the RMASCA-model
}
