% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{getPermutationPValues}
\alias{getPermutationPValues}
\title{Get P values}
\usage{
getPermutationPValues(target, objectlist)
}
\arguments{
\item{target}{The full-model ALASCA object}

\item{objectlist}{List of permutated ALASCA objects}
}
\value{
An ALASCA object
}
\description{
This function calculates P values
}
