% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getPlotLinetypes}
\alias{getPlotLinetypes}
\title{Get linetypes}
\usage{
getPlotLinetypes(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
A list with linetypes
}
\description{
This function returns a list with linetypes for plotting
}
