% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getPlotPalette}
\alias{getPlotPalette}
\title{Get color palette}
\usage{
getPlotPalette(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
A list with colors
}
\description{
This function returns a list with colors for plotting
}
