% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{getRegressionCoefficients}
\alias{getRegressionCoefficients}
\title{Get regression coefficients}
\usage{
getRegressionCoefficients(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function extract the regression coefficients for the ALASCA model
}
