% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{getRegressionPredictions}
\alias{getRegressionPredictions}
\title{Validate underlying regression models}
\usage{
getRegressionPredictions(object)
}
\arguments{
\item{object}{An ALASCA object}

\item{variable}{Variable names for which models to validate (default \code{NA} for all)}
}
\value{
An ALASCA object
}
\description{
This function extract percentiles during validation
}
