% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{getRelevantPCs}
\alias{getRelevantPCs}
\title{Get relevant PCs}
\usage{
getRelevantPCs(object, x)
}
\arguments{
\item{x}{Explanatory power of PC}
}
\value{
A vector with relevant PCs
}
\description{
This function extract percentiles during validation
}
