% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{getScaleFun}
\alias{getScaleFun}
\title{Get a scaling function}
\usage{
getScaleFun(scaleFun_string)
}
\arguments{
\item{scaleFun_string}{String to define scaing function: \code{sdall}, \code{sdref}, \code{sdt1}, \code{sdreft1}}
}
\value{
An ALASCA object
}
\description{
Return scaling function
}
