% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{getValidationPercentilesLoading}
\alias{getValidationPercentilesLoading}
\title{Extract percentiles for loading}
\usage{
getValidationPercentilesLoading(object, objectlist)
}
\arguments{
\item{object}{ALASCA object}

\item{objectlist}{List of ALASCA objects}
}
\value{
An ALASCA object
}
\description{
This function extract percentiles during validation
}
