% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ALASCA}
\alias{plot.ALASCA}
\title{Get an ALASCA object}
\usage{
\method{plot}{ALASCA}(object, component = 1, ...)
}
\arguments{
\item{object}{An \code{\link[=ALASCA]{ALASCA()}} object}

\item{component}{Integer stating which component to return (1 is default)}
}
\value{
An ggplot2 object (or a list og ggplot objects)
}
\description{
This function plots your ALASCA model
}
\examples{
load("PE.Rdata")
plot(model.val)
plot(model, component = "PC2")
plot(model, only = "score", effect = "time")
plot(model, too_dense = 5)
plot(model, highlight = c("PlGF", "IL-1b", "IL-6"))
}
