% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.RMASCA}
\alias{plot.RMASCA}
\title{Get an RMASCA object}
\usage{
\method{plot}{RMASCA}(
  object,
  component = "PC1",
  effect = "both",
  decreasingLoadings = TRUE,
  only = "both",
  enlist = FALSE,
  tooDense = NA
)
}
\arguments{
\item{object}{An RMASCA object}

\item{component}{String stating which component to return (PC1 is default)}

\item{effect}{String stating which effect to return; \code{time}, \code{group}, \code{both} (default)}

\item{decreasingLoadings}{Sort the loadings in dedreasing (default) or increasing order}

\item{only}{String stating which plot to return; \code{both} (default), \code{score} or \code{loading}}

\item{enlist}{Logical. If TRUE, the plots are returned as a list and not as a composed figure (default)}

\item{tooDense}{Numerical. If > 0, only name this number of covariables}
}
\value{
An ggplot2 object (or a list og ggplot objects)
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
