% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{plot_covariate}
\alias{plot_covariate}
\title{Plot covariate coefficients}
\usage{
plot_covariate(...)
}
\arguments{
\item{inheritParams}{plot_covars}
}
\value{
A ggplot2 objects\.
}
\description{
This function returns a plot of the regression coefficients for covariates that is not included in the ASCA model itself
}
