% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R, R/plot.R
\name{plot_covariates}
\alias{plot_covariates}
\title{Plot covariate coefficients}
\usage{
plot_covariates(...)

plot_covariates(...)
}
\arguments{
\item{inheritParams}{plot_covars}
}
\value{
A ggplot2 objects\.

A ggplot2 objects\.
}
\description{
This function returns a plot of the regression coefficients for covariates that is not included in the ASCA model itself

This function returns a plot of the regression coefficients for covariates that is not included in the ASCA model itself
}
