% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_covars}
\alias{plot_covars}
\title{Plot covariate coefficients}
\usage{
plot_covars(
  object,
  covar = NA,
  x_label = NA,
  variables = NA,
  n_limit = 0,
  filename = "covars",
  return_data = FALSE,
  filetype = NA,
  figsize = NA,
  figunit = NA,
  my_theme = NA,
  pvalue = "star"
)
}
\arguments{
\item{object}{An ALASCA object}

\item{covar}{Which covariable(s) to plot (default: \code{NA} which prints all)}

\item{x_label}{Alternative names for the covariables}

\item{return_data}{Set to \code{TRUE} to return data instead of plot}

\item{filetype}{Which filetype you want to save the figure to}

\item{figsize}{A vector containing \code{c(widht,height,dpi)} (default: \code{c(120, 80, 300)})}

\item{figunit}{= "mm",}

\item{my_theme}{A ggplot2 theme}
}
\value{
A ggplot2 objects\.
}
\description{
This function returns a plot of the regression coefficients for covariates that is not included in the ASCA model itself
}
