% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_histogram_score}
\alias{plot_histogram_score}
\title{Plot histogram of validation}
\usage{
plot_histogram_score(
  object,
  component = 1,
  bins = object$n_validation_runs/10,
  effect = "time"
)
}
\arguments{
\item{object}{An ALASCA object}

\item{variable}{The name of the variable(s) you want. \code{NA} returns all (default).}

\item{plottitle}{If \code{TRUE} (default), include variable name as title}
}
\value{
A list of ggplot2 objects per variable
}
\description{
This function plots the validation runs as histograms
}
