% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_participants}
\alias{plot_participants}
\title{Plot covariate coefficients}
\usage{
plot_participants(...)
}
\arguments{
\item{inheritParams}{plot_parts}
}
\value{
A ggplot2 object
}
\description{
This function returns a plot of the regression coefficients for covariates that is not included in the ASCA model itself
}
