% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{plot_parts}
\alias{plot_parts}
\title{Plot participants}
\usage{
plot_parts(
  object,
  variables = NA,
  participant_column = FALSE,
  x_label = NA,
  group_label = NA,
  valueColumn = FALSE,
  timeColumn = "time",
  addSmooth = "loess",
  filename = NA,
  filetype = object$plot.filetype,
  figunit = object$plot.figunit,
  plot.ylabel = "value",
  as.list = FALSE,
  figsize = object$plot.figsize,
  my_theme = object$plot.my_theme
)
}
\arguments{
\item{object}{An ALASCA object or a data frame. If a data frame, you need to specify the column names for participant and value. This also applies if you have not specified the participant column in the ALASCA model before.}

\item{participant_column}{Specify the column with participant identifier. Not necessary if you have already provided it to the ALASCA object}

\item{valueColumn}{Specify column with values (y axis). Not necessary to provide if you are plotting an ALASCA object.}

\item{timeColumn}{Specify column with times (x axis). Defaults to \code{time}.}

\item{filetype}{Which filetype you want to save the figure to}

\item{figsize}{A vector containing \code{c(width,height,dpi)} (default: \code{c(120, 80, 300)})}

\item{my_theme}{A ggplot2 theme to use, defaults to \code{ggplot2::theme_bw()}}

\item{variable}{List of variable names to print. If \code{NA}, return all (default).}

\item{addSmooth.}{Specify which geom_smooth model you want to apply, eg. \code{lm}, \code{glm}, \code{gam}, \code{loess} (default). Set to \code{NA} to remove.}
}
\value{
A list with ggplot2 objects.
}
\description{
This function returns the scores for an ALASCA model
}
