% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{plot_projection}
\alias{plot_projection}
\title{Plot projection of participants}
\usage{
plot_projection(
  object,
  comp = c(1, 2),
  return_data = FALSE,
  filename = NA,
  filetype = NA,
  figsize = NA,
  figunit = NA,
  myTheme = NA
)
}
\arguments{
\item{object}{An ALASCA object}

\item{comp}{Which two components to plot (default: \verb{c(1, 2})}

\item{return_data}{Set to \code{TRUE} to return data instead of plot}

\item{filetype}{Which filetype you want to save the figure to}

\item{figsize}{A vector containing \code{c(widht,height,dpi)} (default: \code{c(120, 80, 300)})}

\item{myTheme}{A ggplot2 theme to use}
}
\value{
A ggplot2 objects.
}
\description{
This function returns a plot of...
}
