% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_val}
\alias{plot_val}
\title{Plot covariate coefficients}
\usage{
plot_val(...)
}
\arguments{
\item{inheritParams}{plot_validation}
}
\value{
A ggplot2 object.
}
\description{
This function returns a plot of the regression coefficients for covariates that is not included in the ASCA model itself
}
