% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotComponents}
\alias{plotComponents}
\title{Plot PCs}
\usage{
plotComponents(
  object,
  comps = c(1, 2),
  filename = NA,
  filetype = NA,
  figsize = NA,
  figunit = NA,
  ...
)
}
\arguments{
\item{object}{An ALASCA object}

\item{filetype}{Which filetype you want to save the figure to}

\item{figsize}{A vector containing \code{c(widht,height,dpi)} (default: \code{c(120, 80, 300)})}

\item{figunit}{= "mm"}

\item{covar}{Which covariable(s) to plot (default: \code{NA} which prints all)}

\item{xlab}{Alternative names for the covariables}

\item{plottext}{If \code{TRUE}, plot time point as text}

\item{validationshape}{Either \code{NA}, "ellipse", or "cross"}

\item{validationlevel}{Defaults to 0.95}

\item{return_data}{Set to \code{TRUE} to return data instead of plot}

\item{myTheme}{A ggplot2 theme to use, defaults to \code{ggplot2::theme_bw()}}
}
\value{
A ggplot2 object.
}
\description{
This function returns a plot of ...
}
