% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCovar}
\alias{plotCovar}
\title{Plot validations models}
\usage{
plotCovar(
  object,
  covar = NA,
  tlab = NA,
  return_data = FALSE,
  myTheme = ggplot2::theme_bw()
)
}
\arguments{
\item{object}{An ALASCA object}

\item{tlab}{Alternative names for the covariables}

\item{return_data}{Set to \code{TRUE} to return data instead of plot}

\item{myTheme}{A ggplot2 theme to use, defaults to \code{ggplot2::theme_bw()}}

\item{component}{Which covariable(s) to plot (default: \code{NA} which prints all)}
}
\value{
A ggplot2 objects\.
}
\description{
This function returns a plot of the validation models
}
