% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotParts}
\alias{plotParts}
\title{Plot participants}
\usage{
plotParts(
  object,
  variable = NA,
  participantColumn = FALSE,
  valueColumn = FALSE,
  timeColumn = "time",
  addSmooth = "loess"
)
}
\arguments{
\item{object}{An RMASCA object or a data frame. If a data frame, you need to specify the column names for participant and value. This also applies if you have not specified the participant column in the RMASCA model before.}

\item{variable}{List of variable names to print. If \code{NA}, return all (default).}

\item{participantColumn}{Specify the column with participant identifier. Not necessary if you have already provided it to the RMASCA object}

\item{valueColumn}{Specify column with values (y axis). Not necessary to provide if you are plotting an RMASCA object.}

\item{timeColumn}{Specify column with times (x axis). Defaults to \code{time}.}

\item{addSmooth.}{Specify which geom_smooth model you want to apply, eg. "\code{lm}", "\code{glm}", "\code{gam}", "\code{loess}" (default). Set to \code{NA} to remove.}
}
\value{
A list with ggplot2 objects.
}
\description{
This function returns the scores for an RMASCA model
}
\examples{
plotParts(df, variable = "CRP", participantColumn = "ID", valueColumn = "value")
plotParts(model, variable = c("CRP", "IL-6"), participantColumn = "ID")
plotParts(model, variable = "CRP", participantColumn = "ID", timeColumn = "GA")[[1]] + ggplot2::labs(x = "GA")

}
