% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPred}
\alias{plotPred}
\title{Plot model predictions}
\usage{
plotPred(
  object,
  variables = object$variablelist,
  filetype = NA,
  figsize = NA,
  figunit = NA,
  dodgewidth = 0.35,
  plotribbon = TRUE,
  as.list = FALSE,
  plot.ylabel = "value",
  myTheme = object$plot.myTheme
)
}
\arguments{
\item{object}{An ALASCA object or a data frame. If a data frame, you need to specify the column names for participant and value. This also applies if you have not specified the participant column in the ALASCA model before.}

\item{filetype}{Which filetype you want to save the figure to}

\item{figsize}{A vector containing \code{c(widht,height,dpi)} (default: \code{c(120, 80, 300)})}

\item{figunit}{= "mm",}

\item{myTheme}{A ggplot2 theme to use, defaults to \code{ggplot2::theme_bw()}}

\item{variable}{List of variable names to print. If \code{NA}, return all (default).}
}
\value{
A list with ggplot2 objects.
}
\description{
This function returns the scores for an ALASCA model
}
\examples{
load("PE.Rdata")
plotPred(model, variable = "IL-6")[[1]] + ggplot2::theme_bw()
do.call(
  ggpubr::ggarrange,
  c(plotPred(model, variable = c("PlGF", "IL-6", "IL-1b", "IFN-g", "Eotaxin-2", "Eotaxin")),
    common.legend = TRUE, legend = "bottom"
  )
)
}
