% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotProjection}
\alias{plotProjection}
\title{Plot projection of participants}
\usage{
plotProjection(
  object,
  comp = c(1, 2),
  return_data = FALSE,
  filetypes = "png",
  figsize = c(12, 8, 300),
  myTheme = ggplot2::theme_bw()
)
}
\arguments{
\item{object}{An ALASCA object}

\item{comp}{Which two components to plot (default: \verb{c(1, 2})}

\item{return_data}{Set to \code{TRUE} to return data instead of plot}

\item{filetypes}{Which filetypes you want to save the figure to}

\item{figsize}{A vector containing \code{c(widht,height,dpi)} (default: \code{c(12, 8, 300)})}

\item{myTheme}{A ggplot2 theme to use, defaults to \code{ggplot2::theme_bw()}}
}
\value{
A ggplot2 objects.
}
\description{
This function returns a plot of...
}
