% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Plot residuals}
\usage{
plotResiduals(
  object,
  variable = NA,
  plottitle = TRUE,
  my_theme = ggplot2::theme_classic()
)
}
\arguments{
\item{object}{An ALASCA object}

\item{variable}{The name of the variable(s) you want. \code{NA} returns all (default).}

\item{plottitle}{If \code{TRUE} (default), include variable name as title}
}
\value{
A list of ggplot2 objects per variable
}
\description{
This function plots the residuals of the underlying linear mixed models
}
