% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotVal}
\alias{plotVal}
\title{Plot validations models}
\usage{
plotVal(
  object,
  component = 1,
  filetype = "png",
  figsize = c(120, 80, 300),
  figunit = "mm",
  myTheme = ggplot2::theme_bw()
)
}
\arguments{
\item{object}{A validated ALASCA object}

\item{component}{Which component to plot (default: 1)}

\item{filetype}{Which filetype you want to save the figure to}

\item{figsize}{A vector containing \code{c(widht,height,dpi)} (default: \code{c(120, 80, 300)})}

\item{myTheme}{A ggplot2 theme to use, defaults to \code{ggplot2::theme_bw()}}
}
\value{
A list with ggplot2 objects.
}
\description{
This function returns a plot of the validation models
}
