% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{predict_scores}
\alias{predict_scores}
\title{Calculate scores for a new data set}
\usage{
predict_scores(data, effect_i = 1, component = 1)
}
\arguments{
\item{data}{A new data set}

\item{effect_i}{The effect to look at}

\item{component}{The PCA components to look at}
}
\value{
A data table with new scores
}
\description{
Calculate scores for a new data set
}
