% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{prepare_validation_run}
\alias{prepare_validation_run}
\title{Make a single validation run}
\usage{
prepare_validation_run(runN = NA)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function ...
}
