% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{prepateValidationRun}
\alias{prepateValidationRun}
\title{Make a single validation run}
\usage{
prepateValidationRun(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function ...
}
