% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{reduce_dimensions}
\alias{reduce_dimensions}
\title{Perform PCA for Limm-PCA}
\usage{
reduce_dimensions(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function performs PCA before "the real PCA"
}
