% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{remove_embedded_data}
\alias{remove_embedded_data}
\title{Remove df from object}
\usage{
remove_embedded_data(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function removes unnecessary data
}
