% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{removeCovars}
\alias{removeCovars}
\title{Remove unwanted covariables}
\usage{
removeCovars(object)
}
\arguments{
\item{object}{An ALASCA object to be sanitized}
}
\value{
An ALASCA object
}
\description{
This function removes coefficients that we do not want in our PCA
}
