% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{rename_columns_to_standard}
\alias{rename_columns_to_standard}
\title{Check if columns are missing}
\usage{
rename_columns_to_standard(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\description{
...
}
