% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{replace_term_in_formula}
\alias{replace_term_in_formula}
\title{Check if response variables are missing}
\usage{
replace_term_in_formula(object, old_term, new_term)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
...
}
