% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{residuals.AlascaModel}
\alias{residuals.AlascaModel}
\title{Prints a marco object}
\usage{
\method{residuals}{AlascaModel}(object, variable = NULL)
}
\arguments{
\item{object}{A marco object.}

\item{variable}{A marco object.}
}
\value{
A data table with residuals
}
\description{
Prints a marco object
}
