% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{residuals.RMASCA}
\alias{residuals.RMASCA}
\title{Get residuals}
\usage{
\method{residuals}{RMASCA}(object, variable = NA)
}
\arguments{
\item{object}{An RMASCA object}

\item{variable}{The name of the variable(s) you want. \code{NA} returns all (default).}
}
\value{
A list of residuals per variable
}
\description{
This function returns the residual of the underlying linear mixed models
}
\examples{
load("PE.Rdata")
residuals(model, variable = c("IL-6", "PlGF"))

}
