% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMASCA.R
\name{RMASCA}
\alias{RMASCA}
\title{Get an RMASCA object}
\usage{
RMASCA(df, formula, separateTimeAndGroup = TRUE, pAdjustMethod = "BH")
}
\arguments{
\item{df}{Data Frame to be analyzed}

\item{formula}{Regression model}

\item{separateTimeAndGroup}{Logical: should time and group effect be separated?}

\item{pAdjustMethod}{Method for correcting p values for multiple testing}
}
\value{
An RMASCA object
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
