% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R, R/validation.R
\name{rotate_matrix_optimize_score}
\alias{rotate_matrix_optimize_score}
\title{Rotate PCA}
\usage{
rotate_matrix_optimize_score(object, target)

rotate_matrix_optimize_score(object, target)
}
\arguments{
\item{object}{ALASCA object to be rotated (and returned)}

\item{target}{ALASCA object acting as target}
}
\value{
An ALASCA object

An ALASCA object
}
\description{
This function rotates loadings and scores during validation

This function rotates loadings and scores during validation
}
\details{
Optimizes the rotation for lowest possible difference in score

Optimizes the rotation for lowest possible difference in score
}
