% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{rotateMatrix}
\alias{rotateMatrix}
\title{Rotate PCA}
\usage{
rotateMatrix(object, target)
}
\arguments{
\item{object}{RMASCA object to be rotated (and returned)}

\item{target}{RMASCA object acting as target}
}
\value{
An RMASCA object
}
\description{
This function calculates the LMM coefficients for the RMASCA-model
}
