% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{rotateMatrixOptimizeScore}
\alias{rotateMatrixOptimizeScore}
\title{Rotate PCA}
\usage{
rotateMatrixOptimizeScore(object, target)
}
\arguments{
\item{object}{ALASCA object to be rotated (and returned)}

\item{target}{ALASCA object acting as target}
}
\value{
An ALASCA object
}
\description{
This function rotates loadings and scores during validation
}
\details{
Optimizes the rotation for lowest possible difference in score
}
