% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R, R/buildModel.R
\name{run_regression}
\alias{run_regression}
\title{Run regressions}
\usage{
run_regression()

run_regression()
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object

An ALASCA object
}
\description{
This function runs the underlying regression models

This function runs the underlying regression models
}
