% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{runRegression}
\alias{runRegression}
\title{Run regressions}
\usage{
runRegression(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function runs the underlying regression models
}
