% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{sanitize_object}
\alias{sanitize_object}
\title{Sanitize an ALASCA object}
\usage{
sanitize_object(object)
}
\arguments{
\item{object}{An ALASCA object to be sanitized}
}
\value{
An ALASCA object
}
\description{
This function checks that the input to an ALASCA object is as expected
}
