% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMASCA.R
\name{sanitizeObject}
\alias{sanitizeObject}
\title{Sanitize an RMASCA object}
\usage{
sanitizeObject(object)
}
\arguments{
\item{object}{An RMASCA object to be sanitized}
}
\value{
An RMASCA object
}
\description{
This function checks that the input to an RMASCA object is as expected
}
