% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R, R/export.R
\name{save_bootstrap_ids}
\alias{save_bootstrap_ids}
\title{Save ALASCA object}
\usage{
save_bootstrap_ids(object)

save_bootstrap_ids(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object

An ALASCA object
}
\description{
Save ALASCA object

Save ALASCA object
}
