% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{save_jackknife_ids}
\alias{save_jackknife_ids}
\title{Save ALASCA object}
\usage{
save_jackknife_ids(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
Save ALASCA object
}
