% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R, R/export.R
\name{save_to_csv}
\alias{save_to_csv}
\title{Save ALASCA object to csv}
\usage{
save_to_csv(
  object,
  filename = NA,
  filepath = NA,
  saveCSV = TRUE,
  saveScores = TRUE,
  saveLoadings = TRUE,
  saveCovars = TRUE,
  csv = "csv",
  ...
)

save_to_csv(
  object,
  filename = NA,
  filepath = NA,
  saveCSV = TRUE,
  saveScores = TRUE,
  saveLoadings = TRUE,
  saveCovars = TRUE,
  csv = "csv",
  ...
)
}
\arguments{
\item{object}{An ALASCA object}
}
\description{
Save ALASCA object to csv

Save ALASCA object to csv
}
