% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R, R/export.R
\name{saveALASCA}
\alias{saveALASCA}
\title{Save ALASCA object}
\usage{
saveALASCA(
  object,
  filename = NA,
  filepath = NA,
  saveCSV = TRUE,
  saveScores = TRUE,
  saveLoadings = TRUE,
  saveCovars = TRUE,
  csv = "csv",
  ...
)

saveALASCA(
  object,
  filename = NA,
  filepath = NA,
  saveCSV = TRUE,
  saveScores = TRUE,
  saveLoadings = TRUE,
  saveCovars = TRUE,
  csv = "csv",
  ...
)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object

An ALASCA object
}
\description{
Save ALASCA object

Save ALASCA object
}
