% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{saveALASCAModel}
\alias{saveALASCAModel}
\title{Save ALASCA object}
\usage{
saveALASCAModel(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
Save ALASCA object
}
