% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{saveJackKnifeID}
\alias{saveJackKnifeID}
\title{Save ALASCA object}
\usage{
saveJackKnifeID(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
Save ALASCA object
}
