% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{savetocsv}
\alias{savetocsv}
\title{Save ALASCA object to csv}
\usage{
savetocsv(
  object,
  filename = NA,
  filepath = NA,
  saveCSV = TRUE,
  saveScores = TRUE,
  saveLoadings = TRUE,
  saveCovars = TRUE,
  csv = "csv",
  ...
)
}
\arguments{
\item{object}{An ALASCA object}
}
\description{
Save ALASCA object to csv
}
