% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{screeplot.ALASCA}
\alias{screeplot.ALASCA}
\title{Get screeplot}
\usage{
\method{screeplot}{ALASCA}(
  object,
  effect = "both",
  filetype = NA,
  figsize = NA,
  figunit = NA,
  myTheme = ggplot2::theme_bw()
)
}
\arguments{
\item{object}{An ALASCA object}

\item{effect}{String stating which effect to return; \code{time}, \code{group}, \code{both} (default)}

\item{filetype}{Which filetype you want to save the figure to}

\item{figsize}{A vector containing \code{c(widht,height,dpi)} (default: \code{c(12, 8, 300)})}

\item{myTheme}{A ggplot2 theme to use, defaults to \code{ggplot2::theme_bw()}}
}
\value{
An ggplot2 object (or a list og ggplot objects)
}
\description{
This function returns a screeplot for an ALASCA model showing what proportion of the variance each component of the model explains
}
\examples{
load("PE.Rdata")
screeplot(model)

}
