% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{screeplot.AlascaModel}
\alias{screeplot.AlascaModel}
\title{Get screeplot}
\usage{
\method{screeplot}{AlascaModel}(model, ...)
}
\arguments{
\item{model}{An ALASCA object}
}
\value{
An ggplot2 object (or a list og ggplot objects)
}
\description{
This function returns a screeplot for an ALASCA model showing what proportion of the variance each component of the model explains
}
