% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{screeplot}
\alias{screeplot}
\title{Get screeplot}
\usage{
screeplot(object, component = 10, ...)
}
\arguments{
\item{object}{An ALASCA object}

\item{component}{The highest principal component to plot}
}
\value{
An ggplot2 object
}
\description{
Get screeplot
}
