% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{screeplot.RMASCA}
\alias{screeplot.RMASCA}
\title{Get screeplot}
\usage{
\method{screeplot}{RMASCA}(object, effect = "both")
}
\arguments{
\item{object}{An RMASCA object}

\item{effect}{String stating which effect to return; \code{time}, \code{group}, \code{both} (default)}
}
\value{
An ggplot2 object (or a list og ggplot objects)
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
