% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R, R/buildModel.R
\name{separate_regression_coefficients}
\alias{separate_regression_coefficients}
\title{Separate time and group effects}
\usage{
separate_regression_coefficients()

separate_regression_coefficients()
}
\arguments{
\item{object}{An ALASCA object to be sanitized}
}
\value{
An ALASCA object

An ALASCA object
}
\description{
This function separates time and group variables if separate_time_and_group = TRUE

This function separates time and group variables if separate_time_and_group = TRUE
}
