% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMM.R
\name{separateLMECoefficients}
\alias{separateLMECoefficients}
\title{Separate time and group effects}
\usage{
separateLMECoefficients(object)
}
\arguments{
\item{object}{An ALASCA object to be sanitized}
}
\value{
An ALASCA object
}
\description{
This function separates time and group variables if separateTimeAndGroup = TRUE
}
