% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{summary.AlascaModel}
\alias{summary.AlascaModel}
\title{Summary}
\usage{
\method{summary}{AlascaModel}(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\description{
Print info about the ALASCA object
}
