% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{switchSign}
\alias{switchSign}
\title{Switch sign}
\usage{
switchSign(value, perc, PCs)
}
\arguments{
\item{value}{Point estimate}

\item{perc}{Uncertainty}

\item{PCs}{Components}
}
\value{
An ALASCA object
}
\description{
This function extract percentiles during validation
}
