% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate}
\alias{validate}
\title{Validate the RMASCA model LMM}
\usage{
validate(object, participantColumn = FALSE)
}
\arguments{
\item{object}{An RMASCA object}
}
\value{
An RMASCA object
}
\description{
This function calculates the LMM coefficients for the RMASCA-model
}
