% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{ALASCA}
\alias{ALASCA}
\title{Get an ALASCA object}
\usage{
ALASCA(df, formula, ...)
}
\arguments{
\item{df}{Data frame to be analyzed}

\item{formula}{Regression model}

\item{separate_time_and_group}{Logical: should time and group effect be separated?}

\item{p_adjust_method}{Method for correcting p values for multiple testing, see p.adjust.methods}

\item{validate}{Logical. If \code{TRUE}, give estimates for robustness}

\item{participant_column}{String. Name of the column containing participant identification}

\item{minimize_object}{Logical. If \code{TRUE}, remove unnecessary clutter, optimize for validation}

\item{scale_function}{Either a custom function or string to define scaling function: \code{sdall}, \code{sdref}, \code{sdt1}, \code{sdreft1}}

\item{scale_function.center}{Boolean. Mean centering as part of scaling}

\item{equal_baseline}{Set to \code{TRUE} (default) to remove interaction between group and first time point}

\item{sum_coding}{Set to \code{TRUE} to use sum coding instead of contrast coding for group (defaults to \code{FALSE})}

\item{plot.x_label}{Defaults to "Time"}

\item{plot.group_label}{Defaults to "Group"}

\item{plot.figsize}{A vector containing \code{c(width,height,dpi)} (default: \code{c(180, 120, 300)})}

\item{plot.figunit}{Defaults to "mm"}

\item{plot.filetype}{Which filetype you want to save the figure to (default: \code{png})}

\item{plot.palette}{List of colors, named by group}

\item{keep_terms}{Additional terms to keep in the model matrix}

\item{stratification_vector}{Vector of same length as \code{df} that specifies stratification groups during validate_ Defaults to \code{NA}, where the group column is used.}

\item{validate_regression}{Whether to validate regression predictions or not (only if \code{validate} is \code{TRUE})}

\item{do_debug}{Print what happens (default: \code{FALSE})}

\item{save}{Save models and plots automatically (default: \code{FALSE})}

\item{filename}{File name to save model and plots (when \code{save = TRUE})}

\item{method}{Defaults to \code{NA} where method is either LM or LMM, depending on whether your formula contains a random effect or not}

\item{use_Rfast}{Boolean. Defaults to \code{TRUE}}

\item{n_validation_folds}{Partitions when validating}

\item{n_validation_runs}{number of validation runs}

\item{validation_method}{among  \code{loo} (leave-one-out, default)}

\item{validation_object}{Don't worry about me:)}

\item{validation_participants}{Don't worry about me:)}
}
\value{
An ALASCA object
}
\description{
\code{ALASCA} initializes an ALASCA model and returns an ALASCA object
}
\details{
This function builds your ALASCA model. It needs a data frame containing at least a column identifying participants, a column called \code{time} contining time information, a column \code{group} containing group information, a column \code{variable} containing variable names, and a value column. In addition you need to specify the model you want, and whether you want to separate group and time effects (defaults to \code{TRUE}).
}
\examples{
load("PE.Rdata")
model <- ALASCA(df = df, formula = value ~ time * group + (1 | ID))
}
