% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{ALASCA}
\alias{ALASCA}
\title{Get an ALASCA object}
\usage{
ALASCA(
  df,
  formula,
  separateTimeAndGroup = TRUE,
  pAdjustMethod = NA,
  participantColumn = "ID",
  validate = FALSE,
  scaleFun = TRUE,
  forceEqualBaseline = FALSE,
  useSumCoding = FALSE,
  method = NA,
  stratificationVector = NA,
  minimizeObject = FALSE,
  plot.xlabel = "Time",
  doDebug = FALSE,
  nValFold = 7,
  nValRuns = 50,
  optimizeScore = TRUE,
  validateRegression = FALSE,
  validationMethod = "loo",
  validationObject = NA,
  validationParticipants = NA
)
}
\arguments{
\item{df}{Data frame to be analyzed}

\item{formula}{Regression model}

\item{separateTimeAndGroup}{Logical: should time and group effect be separated?}

\item{pAdjustMethod}{Method for correcting p values for multiple testing, see p.adjust.methods}

\item{participantColumn}{String. Name of the column containing participant identification}

\item{validate}{Logical. If \code{TRUE}, give estimates for robustness}

\item{scaleFun}{If \code{TRUE} (default), each variable is scaled to unit SD and zero mean. If \code{FALSE}, no scaling is performed. You can also provide a custom scaling function that has the data frame \code{df} as input and output}

\item{forceEqualBaseline}{Set to \code{TRUE} (default) to remove interaction between group and first time point}

\item{useSumCoding}{Set to \code{TRUE} to use sum coding instead of contrast coding for group (defaults to \code{FALSE})}

\item{method}{Defaults to \code{NA} where method is either LM or LMM or Rfast, depending on whether your formula contains a random effect or not}

\item{stratificationVector}{Vector of same length as \code{df} that specifies stratification groups during validation. Defaults to \code{NA}, where the group column is used.}

\item{minimizeObject}{Logical. If \code{TRUE}, remove unnecessary clutter, optimize for validation}

\item{plot.xlabel}{Defaults to "Time"}

\item{doDebug}{Print what happens (default: \code{FALSE})}

\item{nValFold}{Partitions when validating}

\item{nValRuns}{number of validation runs}

\item{validateRegression}{Whether to validate regression predictions or not (only if \code{validate} is \code{TRUE})}

\item{validationMethod}{among  \code{loo} (leave-one-out, default)}

\item{validationObject}{Don't worry about me:)}

\item{validationParticipants}{Don't worry about me:)}
}
\value{
An ALASCA object
}
\description{
\code{ALASCA} initializes an ALASCA model and returns an ALASCA object
}
\details{
This function builds your ALASCA model. It needs a data frame containing at least a column identifying participants, a column called \code{time} contining time information, a column \code{group} containing group information, a column \code{variable} containing variable names, and a value column. In addition you need to specify the model you want, and whether you want to separate group and time effects (defaults to \code{TRUE}).
}
\examples{
load("PE.Rdata")
model <- ALASCA(df = df, formula = value~time*group + (1|ID))

}
