% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{ALASCA}
\alias{ALASCA}
\title{Get an ALASCA object}
\usage{
ALASCA(
  df,
  formula,
  separateTimeAndGroup = TRUE,
  pAdjustMethod = NA,
  participantColumn = "ID",
  validate = FALSE,
  scaleFun = "sdall",
  forceEqualBaseline = FALSE,
  useSumCoding = FALSE,
  method = NA,
  useRfast = TRUE,
  stratificationColumn = "group",
  stratificationVector = NA,
  minimizeObject = FALSE,
  limitsCI = c(0.025, 0.975),
  plot.xlabel = "Time",
  plot.grouplabel = "Group",
  plot.figsize = c(180, 120, 300),
  plot.figunit = "mm",
  plot.filetype = "png",
  plot.palette = NA,
  plot.loadinggroupcolumn = NA,
  plot.loadinggrouplabel = "Variable group",
  plot.palette.end = 0.8,
  explanatorylimit = 0.05,
  doDebug = FALSE,
  nValFold = 7,
  nValRuns = 1000,
  validationQuantileMethod = 2,
  plot.myTheme = ggplot2::theme_classic(),
  keepTerms = c(""),
  save = FALSE,
  validation = FALSE,
  filename = NA,
  filepath = NA,
  limm.nComps = 100,
  lowerLimit = NA,
  savetodisk = FALSE,
  optimizeScore = TRUE,
  validateRegression = TRUE,
  validationMethod = "bootstrap",
  validationIDs = NA,
  validationObject = NA,
  validationAssignNewID = TRUE,
  validationParticipants = NA
)
}
\arguments{
\item{df}{Data frame to be analyzed}

\item{formula}{Regression model}

\item{separateTimeAndGroup}{Logical: should time and group effect be separated?}

\item{pAdjustMethod}{Method for correcting p values for multiple testing, see p.adjust.methods}

\item{participantColumn}{String. Name of the column containing participant identification}

\item{validate}{Logical. If \code{TRUE}, give estimates for robustness}

\item{scaleFun}{If \code{TRUE} (default), each variable is scaled to unit SD and zero mean. If \code{FALSE}, no scaling is performed. You can also provide a custom scaling function that has the data frame \code{df} as input and output}

\item{forceEqualBaseline}{Set to \code{TRUE} (default) to remove interaction between group and first time point}

\item{useSumCoding}{Set to \code{TRUE} to use sum coding instead of contrast coding for group (defaults to \code{FALSE})}

\item{method}{Defaults to \code{NA} where method is either LM or LMM, depending on whether your formula contains a random effect or not. Set to KM or KMM for survival analysis}

\item{useRfast}{Boolean. Defaults to \code{TRUE}}

\item{stratificationVector}{Vector of same length as \code{df} that specifies stratification groups during validation. Defaults to \code{NA}, where the group column is used.}

\item{minimizeObject}{Logical. If \code{TRUE}, remove unnecessary clutter, optimize for validation}

\item{plot.xlabel}{Defaults to "Time"}

\item{plot.grouplabel}{Defaults to "Group"}

\item{plot.figsize}{A vector containing \code{c(width,height,dpi)} (default: \code{c(180, 120, 300)})}

\item{plot.figunit}{Defaults to "mm"}

\item{plot.filetype}{Which filetype you want to save the figure to (default: \code{png})}

\item{plot.palette}{List of colors, named by group}

\item{doDebug}{Print what happens (default: \code{FALSE})}

\item{nValFold}{Partitions when validating}

\item{nValRuns}{number of validation runs}

\item{keepTerms}{Additional terms to keep in the model matrix}

\item{save}{Save models and plots automatically (default: \code{FALSE})}

\item{filename}{File name to save model and plots (when \code{save = TRUE})}

\item{lowerLimit}{A data frame with lower limits for every variable}

\item{validateRegression}{Whether to validate regression predictions or not (only if \code{validate} is \code{TRUE})}

\item{validationMethod}{among  \code{loo} (leave-one-out, default)}

\item{validationObject}{Don't worry about me:)}

\item{validationParticipants}{Don't worry about me:)}
}
\value{
An ALASCA object
}
\description{
\code{ALASCA} initializes an ALASCA model and returns an ALASCA object
}
\details{
This function builds your ALASCA model. It needs a data frame containing at least a column identifying participants, a column called \code{time} contining time information, a column \code{group} containing group information, a column \code{variable} containing variable names, and a value column. In addition you need to specify the model you want, and whether you want to separate group and time effects (defaults to \code{TRUE}).
}
\examples{
load("PE.Rdata")
model <- ALASCA(df = df, formula = value ~ time * group + (1 | ID))
}
