% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R, R/pca.R
\name{do_reduce_dimensions}
\alias{do_reduce_dimensions}
\title{Perform PCA for Limm-PCA}
\usage{
do_reduce_dimensions()

do_reduce_dimensions()
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object

An ALASCA object
}
\description{
This function performs PCA before "the real PCA"

This function performs PCA before "the real PCA"
}
