% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMASCA.R
\name{flipIt}
\alias{flipIt}
\title{Flip an RMASCA object}
\usage{
flipIt(object)
}
\arguments{
\item{object}{An RMASCA object}
}
\value{
An RMASCA object
}
\description{
Changes the sign of loadings and scores
}
