% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{flipIt}
\alias{flipIt}
\title{Flip an ALASCA object}
\usage{
flipIt(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
Changes the sign of loadings and scores
}
