% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_covars}
\alias{get_covars}
\title{Get coefficients for covariates}
\usage{
get_covars(object, ...)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
A data table with regression coefficients
}
\description{
Get coefficients for covariates
}
