% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_effect_matrix}
\alias{get_effect_matrix}
\title{Get effect matrix}
\usage{
get_effect_matrix()
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function calculates the effect matrix
}
