% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_loadings}
\alias{get_loadings}
\title{Returns model scores}
\usage{
get_loadings(object, ...)
}
\arguments{
\item{object}{An ALASCA object.}
}
\value{
A data table with loadings
}
\description{
Returns model scores
}
