% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_scaling_function}
\alias{get_scaling_function}
\title{Get a scaling function}
\usage{
get_scaling_function(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
Return scaling function
}
