% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_scores}
\alias{get_scores}
\title{Get scores}
\usage{
get_scores(object, component = 0)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
A list with scores for time (and group), and the exploratory power for each component
}
\description{
This function returns the scores for an ALASCA model
}
