% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_validation_percentiles}
\alias{get_validation_percentiles}
\title{Extract percentiles}
\usage{
get_validation_percentiles(objectlist)
}
\arguments{
\item{objectlist}{List of ALASCA objects}
}
\description{
This function extract percentiles during validation
}
