% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{get_validation_percentiles_covars}
\alias{get_validation_percentiles_covars}
\title{Extract percentiles for covariates}
\usage{
get_validation_percentiles_covars(object, objectlist)
}
\arguments{
\item{object}{ALASCA object}

\item{objectlist}{List of ALASCA objects}
}
\value{
An ALASCA object
}
\description{
This function extract percentiles for validation of covariates
}
