% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{get_validation_percentiles_score}
\alias{get_validation_percentiles_score}
\title{Extract percentiles for scores}
\usage{
get_validation_percentiles_score(objectlist)
}
\arguments{
\item{objectlist}{List of ALASCA objects}
}
\description{
This function extract percentiles during validation of scores
}
