% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getCovars}
\alias{getCovars}
\title{Get covariables}
\usage{
getCovars(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
A list with scores for time (and group), and the exploratory power for each component
}
\description{
This function returns the other covariables in an ALASCA model
}
