% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMM.R
\name{getEffectMatrix}
\alias{getEffectMatrix}
\title{Get effect matrix}
\usage{
getEffectMatrix(object)
}
\arguments{
\item{object}{An RMASCA object}
}
\value{
An RMASCA object
}
\description{
This function separates time and group variables if separateTimeAndGroup = TRUE
}
