% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getLoadingPlot}
\alias{getLoadingPlot}
\title{Get loading plot}
\usage{
getLoadingPlot(
  object,
  component = "PC1",
  effect = "time",
  decreasingLoadings = TRUE,
  tooDense = NA,
  highlight = NA
)
}
\arguments{
\item{object}{An RMASCA object}

\item{component}{Which component to plot?}

\item{effect}{Plot time or group}

\item{decreasingLoadings}{Logical. Should loading sbe sorted in decreasing order?}
}
\value{
A ggplot object
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
