% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getLoadings}
\alias{getLoadings}
\title{Get loadings}
\usage{
getLoadings(object)
}
\arguments{
\item{object}{An RMASCA object}
}
\value{
A list with loadings for time (and group), and the exploratory power for each component
}
\description{
This function  returns the loadings for an RMASCA model
}
