% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{getRegressionPredictions}
\alias{getRegressionPredictions}
\title{Validate underlying regression models}
\usage{
getRegressionPredictions(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function calcuates predictions from each regression model
}
