% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getScorePlot}
\alias{getScorePlot}
\title{Get score plot}
\usage{
getScorePlot(
  object,
  component = 1,
  effect = "time",
  filetype = NA,
  figsize = NA,
  figunit = NA,
  plotribbon = TRUE,
  dodgewidth = 0.35,
  myTheme = ggplot2::theme_bw()
)
}
\arguments{
\item{object}{An ALASCA object}

\item{component}{Which component to plot?}

\item{effect}{Plot time or group}

\item{filetype}{Which filetype you want to save the figure to}

\item{figsize}{A vector containing \code{c(width,height,dpi)} (default: \code{c(120, 80, 300)})}

\item{myTheme}{A ggplot2 theme to use, defaults to \code{ggplot2::scale_color_viridis_d(end = 0.9) + ggplot2::theme_bw()}}
}
\value{
A ggplot object
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
