% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getScores}
\alias{getScores}
\title{Get scores}
\usage{
getScores(object)
}
\arguments{
\item{object}{An RMASCA object}
}
\value{
A list with scores for time (and group), and the exploratory power for each component
}
\description{
This function returns the scores for an RMASCA model
}
