% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{getValidationPercentiles}
\alias{getValidationPercentiles}
\title{Extract percentiles}
\usage{
getValidationPercentiles(object, objectlist)
}
\arguments{
\item{object}{RMASCA object}

\item{objectlist}{List of RMASCA objects}
}
\value{
An RMASCA object
}
\description{
This function calculates the LMM coefficients for the RMASCA-model
}
