% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ALASCA}
\alias{plot.ALASCA}
\title{Get an ALASCA object}
\usage{
\method{plot}{ALASCA}(
  object,
  component = "PC1",
  effect = "both",
  decreasingLoadings = TRUE,
  only = "both",
  enlist = FALSE,
  tooDense = NA,
  highlight = NA
)
}
\arguments{
\item{object}{An \code{\link[=ALASCA]{ALASCA()}} object}

\item{component}{String stating which component to return ("PC1" is default)}

\item{effect}{String stating which effect to return; \code{time}, \code{group}, \code{both} (default)}

\item{decreasingLoadings}{Sort the loadings in decreasing (default) or increasing order}

\item{only}{String stating which plot to return; \code{both} (default), \code{score} or \code{loading}}

\item{enlist}{Logical. If TRUE, the plots are returned as a list and not as a composed figure (default)}

\item{tooDense}{Integer, If > 0, only name this number of covariables}

\item{highlight}{Vector of strings with variables to highlight in the loadings plot}
}
\value{
An ggplot2 object (or a list og ggplot objects)
}
\description{
This function plots your ALASCA model
}
\examples{
load("PE.Rdata")
plot(model.val)
plot(model, component = "PC2")
plot(model, only = "score", effect = "time")
plot(model, tooDense = 5)
plot(model, highlight = c("PlGF", "IL-1b", "IL-6"))

}
