% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotParts}
\alias{plotParts}
\title{Plot participants}
\usage{
plotParts(
  object,
  variable = NA,
  participantColumn = FALSE,
  valueColumn = FALSE,
  timeColumn = "time",
  addSmooth = "loess"
)
}
\arguments{
\item{object}{An RMASCA object or a data frame. If a data frame, you need to specify the column names for participant and value. This also applies if you have not specified the participant column in the RMASCA model before.}

\item{variable}{List of variable names to print. If \code{NA}, return all (default).}

\item{participantColumn}{Specify the column with participant identifier. Not necessary if you have already provided it to the RMASCA object}

\item{valueColumn}{Specify column with values (y axis). Not necessary to provide if you are plotting an RMASCA object.}

\item{timeColumn}{Specify column with times (x axis). Defaults to \code{time}.}

\item{addSmooth.}{Specify which geom_smooth model you want to apply, eg. \code{lm}, \code{glm}, \code{gam}, \code{loess} (default). Set to \code{NA} to remove.}
}
\value{
A list with ggplot2 objects.
}
\description{
This function returns the scores for an RMASCA model
}
\examples{
load("PE.Rdata")
plotParts(df, variable = "IL-6", participantColumn = "ID", valueColumn = "value")
do.call(
    ggpubr::ggarrange,
    c(plotParts(df, participantColumn = "ID", timeColumn = "GA", valueColumn = "value", addSmooth = NA, 
    variable = c("PlGF", "IL-6", "IL-1b", "IFN-g", "Eotaxin-2", "Eotaxin")), 
      common.legend = TRUE, legend = "bottom")
  )
plotParts(model, variable = "IL-6", participantColumn = "ID", timeColumn = "GA")[[1]] + ggplot2::labs(x = "GA")

}
