% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPred}
\alias{plotPred}
\title{Plot model predictions}
\usage{
plotPred(
  object,
  variable = NA,
  filetype = "png",
  figsize = c(120, 80, 300),
  myTheme = ggplot2::theme_bw()
)
}
\arguments{
\item{object}{An ALASCA object or a data frame. If a data frame, you need to specify the column names for participant and value. This also applies if you have not specified the participant column in the ALASCA model before.}

\item{variable}{List of variable names to print. If \code{NA}, return all (default).}

\item{filetype}{Which filetype you want to save the figure to}

\item{figsize}{A vector containing \code{c(widht,height,dpi)} (default: \code{c(120, 80, 300)})
figunit = "mm",}

\item{myTheme}{A ggplot2 theme to use, defaults to \code{ggplot2::theme_bw()}}
}
\value{
A list with ggplot2 objects.
}
\description{
This function returns the scores for an ALASCA model
}
\examples{
load("PE.Rdata")
plotPred(model, variable = "IL-6")[[1]] + ggplot2::theme_bw()
do.call(
  ggpubr::ggarrange,
  c(plotPred(model, variable = c("PlGF", "IL-6", "IL-1b", "IFN-g", "Eotaxin-2", "Eotaxin")), 
    common.legend = TRUE, legend = "bottom")
)
}
