% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotVal}
\alias{plotVal}
\title{Plot validations models}
\usage{
plotVal(object, component = 1, myTheme = ggplot2::theme_bw())
}
\arguments{
\item{object}{A validated ALASCA object}

\item{component}{Which component to plot (default: 1)}

\item{myTheme}{A ggplot2 theme to use, defaults to \code{ggplot2::theme_bw()}}
}
\value{
A list with ggplot2 objects.
}
\description{
This function returns a plot of the validation models
}
