% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotVal}
\alias{plotVal}
\title{Plot validations models}
\usage{
plotVal(object, component = "PC1")
}
\arguments{
\item{object}{A validated RMASCA object}
}
\value{
A list with ggplot2 objects.
}
\description{
This function returns a plot of the validation models
}
