% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{prepareValidationRun}
\alias{prepareValidationRun}
\title{Make a single validation run}
\usage{
prepareValidationRun(object, runN = NA)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function ...
}
