% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMASCA.R
\name{removeEmbedded}
\alias{removeEmbedded}
\title{Remove df from objectt}
\usage{
removeEmbedded(object)
}
\arguments{
\item{object}{An RMASCA object}
}
\value{
An RMASCA object
}
\description{
This function checks that the input to an RMASCA object is as expected
}
