% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{removeEmbedded}
\alias{removeEmbedded}
\title{Remove df from object}
\usage{
removeEmbedded(object)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
}
\description{
This function removes unnecessary data
}
