% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMASCA.R
\name{RMASCA}
\alias{RMASCA}
\title{Get an RMASCA object}
\usage{
RMASCA(
  df,
  formula,
  separateTimeAndGroup = TRUE,
  pAdjustMethod = NA,
  participantColumn = FALSE,
  validate = FALSE,
  scale = TRUE,
  minimizeObject = FALSE,
  nValFold = 7,
  nValRuns = 50,
  validationMethod = "loo",
  validationObject = NA,
  validationParticipants = NA
)
}
\arguments{
\item{df}{Data Frame to be analyzed}

\item{formula}{Regression model}

\item{separateTimeAndGroup}{Logical: should time and group effect be separated?}

\item{pAdjustMethod}{Method for correcting p values for multiple testing, see p.adjust.methods}

\item{participantColumn}{String. Name of the column containing participant identification}

\item{validate}{Logical. If \code{TRUE}, give estimates for robustness}

\item{scale}{If \code{TRUE} (default), each variable is scaled to unit SD and zero mean}

\item{minimizeObject}{Logical. If \code{TRUE}, remove unnecessary clutter, optimize for validation}

\item{nValFold}{Partitions when validating}

\item{nValRuns}{number of validation runs}

\item{validationMethod}{among  \code{loo} (leave-one-out, default)}

\item{validationObject}{Don't worry about me:)}

\item{validationParticipants}{Don't worry about me:)}
}
\value{
An RMASCA object
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
