% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMASCA.R
\name{RMASCA}
\alias{RMASCA}
\title{Get an RMASCA object}
\usage{
RMASCA(
  df,
  formula,
  separateTimeAndGroup = TRUE,
  pAdjustMethod = NA,
  participantColumn = FALSE,
  validate = FALSE,
  scale = TRUE,
  forceEqualBaseline = FALSE,
  minimizeObject = FALSE,
  nValFold = 7,
  nValRuns = 50,
  validationMethod = "loo",
  validationObject = NA,
  validationParticipants = NA
)
}
\arguments{
\item{df}{Data frame to be analyzed}

\item{formula}{Regression model}

\item{separateTimeAndGroup}{Logical: should time and group effect be separated?}

\item{pAdjustMethod}{Method for correcting p values for multiple testing, see p.adjust.methods}

\item{participantColumn}{String. Name of the column containing participant identification}

\item{validate}{Logical. If \code{TRUE}, give estimates for robustness}

\item{scale}{If \code{TRUE} (default), each variable is scaled to unit SD and zero mean}

\item{minimizeObject}{Logical. If \code{TRUE}, remove unnecessary clutter, optimize for validation}

\item{nValFold}{Partitions when validating}

\item{nValRuns}{number of validation runs}

\item{validationMethod}{among  \code{loo} (leave-one-out, default)}

\item{validationObject}{Don't worry about me:)}

\item{validationParticipants}{Don't worry about me:)}
}
\value{
An RMASCA object
}
\description{
\code{RMASCA} initializes an RMASCA model and returns an RMASCA object
}
\details{
This function builds your RMASCA model. It needs a data frame containing at least a column identifying participants, a column called \code{time} contining time information, a column \code{group} containing group information, a column \code{variable} containing variable names, and a value column. In addition you need to specify the model you want, and whether you want to separate group and time effects (defaults to \code{TRUE}).
}
\examples{
load("PE.Rdata")
model <- RMASCA(df = df, formula = value~time*group + (1|ID))

}
