% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{rotate_matrix}
\alias{rotate_matrix}
\title{Rotate PCA}
\usage{
rotate_matrix(target)
}
\arguments{
\item{target}{ALASCA object acting as target}
}
\description{
This function rotates loadings and scores during validation
}
