% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{run_regression}
\alias{run_regression}
\title{Run regressions}
\usage{
run_regression()
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
An ALASCA object
We need to modify the model matrix. Therefore, self[\link{"formula"}][\link{"formula"}] contains the original terms, whereas
self[\link{"formula"}][\link{"regression_formula"}] has replaced the terms with the modified model matrix
We need to modify the model matrix. Therefore, self[\link{"formula"}][\link{"formula"}] contains the original terms, whereas
self[\link{"formula"}][\link{"regression_formula"}] has replaced the terms with the modified model matrix
}
\description{
This function runs the underlying regression models
}
