% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{saveALASCAModel}
\alias{saveALASCAModel}
\title{Save ALASCA object}
\usage{
saveALASCAModel(object, filename = NA, filepath = NA)
}
\arguments{
\item{object}{An ALASCA object}
}
\value{
Full file name of the saved object (String)
}
\description{
Save ALASCA object
}
