% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{saveALASCAPlot}
\alias{saveALASCAPlot}
\title{Save figure}
\usage{
saveALASCAPlot(
  object,
  g,
  filetype = NA,
  figsize = NA,
  prefix = "plot/",
  suffix = "",
  figunit = NA
)
}
\arguments{
\item{g}{ggplot-object}
}
\value{
A ggplot2 objects.
}
\description{
Save figure
}
