% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{saveALASCAPlot}
\alias{saveALASCAPlot}
\title{Save figure}
\usage{
saveALASCAPlot(
  object,
  g,
  filetypes = "png",
  figsize = c(12, 8, 300),
  suffix = ""
)
}
\arguments{
\item{g}{ggplot-object}
}
\value{
A ggplot2 objects.
}
\description{
Save figure
}
