% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALASCA.R
\name{screeplot.AlascaModel}
\alias{screeplot.AlascaModel}
\title{Get screeplot}
\usage{
\method{screeplot}{AlascaModel}(object, component = 10, ...)
}
\arguments{
\item{object}{An ALASCA object}

\item{component}{The highest principal component to plot}
}
\value{
An ggplot2 object
}
\description{
This function returns a screeplot for an ALASCA model showing what proportion of the variance each component of the model explains
}
